% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envirotox-25-acute.R
\docType{data}
\name{envirotox_25_acute}
\alias{envirotox_25_acute}
\title{Acute Species Sensitivity Data as compiled in 2025}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10833 rows and 9 columns.
}
\usage{
envirotox_25_acute
}
\description{
Acute species sensitivity data from Envirotox (Connors et al. 2019)
}
\examples{

head(envirotox_25_acute)

}
\references{
Connors, K.A., Beasley, A., Barron, M.G., Belanger, S.E., Bonnell, M., Brill, J.L., de Zwart, D., Kienzler, A., Krailler, J., Otter, R., Phillips, J.L., and Embry, M.R. 2019. Creation of a Curated Aquatic Toxicology Database: EnviroTox. Enviro Toxic and Chemistry 38(5): 1062–1073. doi:10.1002/etc.4382.
}
\keyword{datasets}
