% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envirotox-chemical.R
\docType{data}
\name{envirotox_chemical}
\alias{envirotox_chemical}
\title{Envirotox Chemical Data}
\format{
\subsection{\code{envirotox_chemical}}{

A data frame with 744 rows and 2 columns:
\describe{
\item{Chemical}{Chemical name (chr)}
\item{OriginalCAS}{Original Chemical Abstracts Service Registry Number (int)}
}
}
}
\source{
\url{https://envirotoxdatabase.org/}
}
\usage{
envirotox_chemical
}
\description{
Envirotox Chemical Data
}
\examples{

head(envirotox_chemical)

}
\keyword{datasets}
