% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-datasets.R
\name{list_datasets}
\alias{list_datasets}
\title{Lists the datasets}
\usage{
list_datasets()
}
\value{
A character vector of the dataset names.
}
\description{
Lists the names of the datasets in the package.
}
\examples{
list_datasets()
}
