% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yanagihara-24-acute.R
\docType{data}
\name{yanagihara_24_acute}
\alias{yanagihara_24_acute}
\title{Acute Species Sensitivity Data as compiled by Yanagihara et al. (2024)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2173 rows and 6 columns.
}
\usage{
yanagihara_24_acute
}
\description{
Acute species sensitivity data from Envirotox as worked up by Yanagihara et al. (2024).
It is important to note that Yanagihara et al. (2024)
exclude datasets with a bimodality coefficient (BC) > 0.555.
}
\details{
The columns are as follows:

\describe{
\item{Chemical}{The chemical name (chr).}
\item{Conc}{The chemical concentration in micrograms per litre (dbl).}
\item{Species}{The species name (chr).}
\item{Type}{'Acute' (chr).}
\item{Group}{Acute (chr).}
\item{Original_CAS}{Original CAS (chr).}
}
}
\examples{

head(yanagihara_24_acute)

}
\keyword{datasets}
