% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envirotox-package.R
\docType{package}
\name{envirotox-package}
\alias{envirotox}
\alias{envirotox-package}
\title{envirotox: Envirotox Datasets}
\description{
The envirotox R data package provides Species Sensitivity Distribution (SSD) datasets from the Envirotox database (Connors et al. 2019). The datasets are provided for assessing general patterns in SSD data and testing code. The datasets should not be used to draw any conclusions about the toxicity of the individual chemicals.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://poissonconsulting.github.io/envirotox/}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@pissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Authors:
\itemize{
  \item Rebecca Fisher \email{R.Fisher@aims.gov.au}
}

Other contributors:
\itemize{
  \item Ayla Pearson \email{ayla@poissonconsulting.ca} (\href{https://orcid.org/0000-0001-7388-1222}{ORCID}) [contributor]
}

}
\keyword{internal}
