% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envirotox-chemical.R
\docType{data}
\name{envirotox_chemical}
\alias{envirotox_chemical}
\title{Species Sensitivity Chemical Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 744 rows and 2 columns.
}
\usage{
envirotox_chemical
}
\description{
The columns are as follows:
}
\details{
\describe{
\item{Chemical}{The chemical name (chr).}
\item{Original_CAS}{Original CAS (chr).}
}
}
\examples{

head(envirotox_chronic)

}
\references{
Connors, K.A., Beasley, A., Barron, M.G., Belanger, S.E., Bonnell, M., Brill, J.L., de Zwart, D., Kienzler, A., Krailler, J., Otter, R., Phillips, J.L., and Embry, M.R. 2019. Creation of a Curated Aquatic Toxicology Database: EnviroTox. Enviro Toxic and Chemistry 38(5): 1062–1073. doi:10.1002/etc.4382.
}
\keyword{datasets}
